Add database file loading manifest.toml
I'm not sure where this will end up in the end, but suffice it to say
that a simple, easy-to-parse-with-libraries format for just identifying
the load order of the different database starting files should suffice
for now.  Eventually, the system level configuration of the database
will either be in Elixir scripts or TOML, so lets start with TOML since
it's simpler than anything.  Maybe we move to an elixir built script
which includes the load order later.