Refactor enums for optons/user_options
Rather than hard code in the options that these enums need to support
at this time, store the various flags and options in a JSON object for
now.  This might be a decent long term strategy, but if not we can
always refactor later when we have a more concrete idea of what we want
to support and how we want to support it.