Add functional_type to enum_contact_types
After reviewing the previous designs on this table, I concluded that
something like the original was probably best for now, especially with
the idea of localized formats for addresses and phones.

In order to accomplish that, this seems to be a case where the old
functional_type column makes sense.  The types are all mutually
exclusive and don't have overlapping functional traits like state can.
So in this case it's a better fit than the options column.

Add the functional_type and establish the expected types.  Do that here
so we have some guide as what to build later in the application.