Starting table of contents file
This is the book introduction point.  It's really a stub as I want to
see how this is going to work with GitHub.  I may need to name the file
as README.md.

(Also consider using the GitHub wiki feature... we'll see... it does
need to be usable portable outside of GitHub).