Update .gitignore for LaTeX based book development
Note that we assume that local build files are in the
documentation/TheBook/_build directory.  This is a bit fragile since the
build system on any one developer's system could be anything, including
locations not covered by these exclusions.

Nonetheless, our expectation is that these are the directions and files
used for building the The Book and developers should adjust their own
build systems respecting this intent.