Add starting document of The Book (#3)
* Starting table of contents file

This is the book introduction point.  It's really a stub as I want to
see how this is going to work with GitHub.  I may need to name the file
as README.md.

(Also consider using the GitHub wiki feature... we'll see... it does
need to be usable portable outside of GitHub).

* Update .gitignore for LaTeX based book development

Note that we assume that local build files are in the
documentation/TheBook/_build directory.  This is a bit fragile since the
build system on any one developer's system could be anything, including
locations not covered by these exclusions.

Nonetheless, our expectation is that these are the directions and files
used for building the The Book and developers should adjust their own
build systems respecting this intent.

* Remove testing markdown file

The primary documentation will be written with LaTeX now as markdown is
too flaky for the job.

* Add The Book starter main.tex

This initial commit includes a reasonable title page, copyright page, a
table of contents, and a file for the logo.

This almost certainly is not where we want to land with this, but the
output isn't embarrassing either.

* Add a placeholder overview chapter to The Book

* Refactor root images directory to be global_images

Each chapter may well have its own images directory, so for those
images that are truly global, like logos, make the nature of the root
images directory explicit in the name.

Also update references to images where they exist in documents.

* Refactor overview directory into chapters/overview

There are global level directories and mixing the chapter directories
with those global directories make it ambiguous as to what purpose each
directory serves.

So create a root level "chapters" directory to explicitly hold all of
the chapter directories.

Co-authored-by: Steven C. Buttgereit <scb@musesystems.com>