Refactor mstr_customer_relationship name
The new name of this table is mstr_entity_selling_entities.  The reason
for the change is that as the use of relationships becomes more
prevalent between entities, persons, and places, the table naming
convention will need to be more clear to know which subject and object
are being related and for what purpose.

Since we're acknowledging the graph-like nature of the CRM data, let's
take inspiration from RDF for this purpose and name the relationship
tables in the form of triples.  This is probably a stretch and the goal
is not to create anything that is formally RDF.