Change The Book to be Word based (#4)
* Remove tex files for The Book

Just going to do the documentation in Word for now.  So get rid of the
LaTex files.  The little side venture in LaTex was worthwhile as we now
know what that is about and it's limitations and such.

* Start of the Word based version of "The Book"

It won't version nicely, but it's here nonetheless.  This is a starter
commit with some generic starting text.

* Remove documentation GitHub Actions

I expect this to come back, but for now, we're not building LaTex docs
anymore.

Co-authored-by: Steven C. Buttgereit <scb@musesystems.com>