Refactor login table name to be more purposeful
The enum_authentication_states was appropriate when the logins table and
the authenticators table were one and the same.  But now that they are
to be split into a login table and an authenticators table, the state
should be made more clearly to apply to the login table.

So rename enum_authentication_states to enum_login_states.