Add owning_entity_id to mstr_contacts
Contact records may be added specifically in support of specific entity
relationships.  For example, a person may have a staffing relationship
with an entity.  The entity that is the employer may want to add office
information as contact information that is pertinent only to that
relationship.

Adding an owning_entity_id field to contextualize the contact for the
person should help make that relationship and management of the contact
information clearer.