Add last_login and last_attempted_login columns
This adds these columns to syst_person_logins.  While the authoritative
data will lie in the syst_interaction_logs, combing those for certain
basic information will be impractical for faster checks.

Put some basic data in the login table.

We may want to expand this later, but for now this will suffice.