Add place_id column to mstr_person_contact_roles
In some instances we want contact information that's also derived from
a place.  Rather than force the recreation of that information, allow it
to be linked to the same information in the place.

There are downsides to this approach, so it may functionally just be a
copy of the place contact record to the a new person oriented contact
record, but recording the place here still gives you a way to tell how
the contact originated.

Ambiguity resulting from this approach will be worked out at a later
date.