Add schemaspy BRM related documentation
This is the initial retrospection into the the BRM module schema and the
first creation of the schemaspy documentation for MuseBMS.  As such
there are a lot of files which are essentially boilerplate being added
here in addition to the actuall application BRM schema specific docs.

Hopefully subsequent commits will be more focused on the specific
changes made.