Add mix release.init boilerplate files
This is just getting the basics of the rel directory in place.  No
changes beyond what mix release.init does is present.

As a note, it's unclear whether or not this directory should be placed
under version control; I couldn't find any statements one way or the
other... will version for now since I know we want our overlays
versioned.