Add Msbms app entry point & initial supervisors
This commit adds the Elixir application entry point for Msbms, the
application server, and initial parts of the supervisor tree which
handle the data access management pieces of the system.

The general idea is that there is a data supervisor which manages a
registry process to find instance specific Repo processes and then a
child supervisor which manages the individual instances.  I do this
because if the registry is not available or the Repo supervisor is not
available then the data supervisor must either restart them or die (and
if the data supervisor itself dies, the app is dead at that point).  The
repo supervisor will try to restart individual repos, but it won't take
down it's sibling repos if the failure is fatal.

-- Add msbms/application.ex file
-- Add msbms/system/data_access_supervisor.ex file
-- Update mix.exs to start the Msbms application via calls to the
   msbms/application.ex module functions as appropriate for the
   Application behaviour.