Add instance management database
This commit add the concept of instance management to the database.  The
gist is that there is a system operations database which tells the
application where instance databases exist so that repos can be built
at application runtime.  This is the starting point for multi-tenancy
support, though I think it will be useful in other scenarios as well,
production vs. sandboxes for example.

  - Update tools/database_script files for a bootstraping change.
    Before this commit, there was a single application database which
    was responsible for both application data and system operations
    configuration.  Now, the system will bootstrap from a dedicated
    system operations configuration database, but that database must
    have some initial setup prior to running the application server.
    This changes the prior reset_dev_database script to be not just a
    dev tool and also separates the tear-down operations (only useful
    for dev) from the bootstrap portion of the code (useful in dev and
    to bootstrap a running application server).

  - Add a database definition for system operations configuration.  This
    is just a start, but I expect to be able to get some distance with
    just this.