Add msbms_startup_options.sample.toml
This is a startup configuration file needed mostly to tell the
application server how to find it's system management database from
which it can load the runtime configuration generally.

There a couple other bits.  The most important of which is the
instance_salt value.  This is used to construct credentials for system
created database user accounts;  doing it this way is probably a bad
idea.  This "salt" combines with an instance specific random set of
bytes which then forms the authenticator for the instance specific
logins of the connection pool.  Ideally, we'd want to have client certs
for the individual instance logins, but that's going to be more work
and for the time being this should be good enough (which is all that you
ever want for security... right?....)