Refactor database code organization
In the end, the duplicated copies of get_exception_details.sql and
trig_b_iu_set_diagnostic_columns.sql got to be too much for me; that
and the other database code needed better organization.

So change the database code directory hierarchy to better reflect what
we're doing now.

Note that it's conceivable that other unrelated changes made their way
into this commit.