Refactor database documentation organization
This change brings the structure in-line with both the new organization
of the database source code as well as introduces the concept that we
have two different databases to run the application.

This organization will like change and the documentation matures into
something of more a manual; for example between users and operators
rather than by the logical organization of the software.... but it's a
start.