Refactor database_migrations table for r/v/u
Previously we were planning to use a simple incrementing sequence or
time based migration numbering scheme.  But really since we're baking in
release/version/update ideas into the app pretty solidly, and also
managing all of it via a mono-repo, it probably should be baked into the
the database versioning strategy as well.

So just make our migrator expect to apply files named for our specific
version.  The final migration naming scheme therefore becomes:

      global.RRRR.VVVV.UUUUUU.eex.ssl
    instance.RRRR.VVVV.UUUUUU.eex.ssl

    R = release digit
    V = version digit
    U = update digit

Zero padding the numbers in the file name is for filesystem sorting
which may be dictionary.

Specific patches will be added to the patch file much the same way the
starter database is built.

Also in this commit: Refactor the manifest files to make them an array
of migration records, one for each release/version/update.  This should
allow us to complete build both the starting database and the updates to
come from the same manifest file.  And using one manifest means only one
reference for finding where sources are kept and their ordering.