Remove the database_migrations seed data file
With the release/version/update refactor, we won't be hacking in the
initial starting database create as a faux migration.  Rather the
migrator will handle cases where there is no database_migrations table
as a regular migration with the default of whatever the starter for the
release/version/0 combination is for the starter; by the time the
migrator will want to update the migrations table with a successful
migration application, the table will exist.