Refactor database_migrations table load order
Move the database_migrations table creation to earlier in the process in
both the global and instance manifests.

While not strictly speaking necessary, it's still a good idea since a
script should be able to query the state of the database_migrations
table.  None in the starter databases do now, but as a matter of good
form make the change anyway.