Add common/misc/no_changes.eex.sql file
Right now the plan is to have a placeholder migration for any update
that doesn't have database changes.  This way we get a positive
indication that the database was considered during any update and we
see all updates in the database_migrations table; we aren't wondering if
any skips are because there was a missing migration or because there was
just no update.