Rename global system configuration sql file
This renames the file conf_instance_management_settings.eex.sql to
conf_global_settings.eex.sql.

The original file name was made prior to the current meanings of global
and instance being well defined for the application.  Now that
"instance" means a single running instance of the user data and that
"global" means the running system overall and components that support
the global system, the new name is more appropriate.  Also, it's more
concise so a better name all around.

Also make the change in the manifest_global.toml file where the old name
was previously mentioned to build it into the migration.