Refactor migration/versioning for sponsor changes
Thinking through how versioning and migrations might evolve over time,
we want to admit the possibility that there will be user sponsored
changes, customizations, or other similar special cases where the system
may need to perform a one-off migration not otherwise made generally
available.

This change does two things to support this approach.  First we add two
more versioning segments.  First is a sponsor ID which identifies the
sponsoring user; even in general releases we always identify some
sponsor so we designate Muse Systems as the sponsor for such generally
released software.  The next versioning segment is the patch that is
specific for the sponsor.  All client sponsor codes should come after
the Muse Systems codes and so should be applied after all standard
release patches.

The second chance is now the migration file names represent version
numbers as Base 36 numbers.   This allows for the file names to be
shorter and more easily read while still being sortable for migration
ordering purposes.  So an example is that Muse Systems's sponsor code
for general release software is "820" in base 10 which in base 36 is
"MS".

All of this and the acceptable ranges for each segment are documented in
the database_migrations.eex.sql file.