Refactor repo.ex into purpose specific modules
This change removes the typical repo.ex used to create an Ecto Repo and
creates two new files that use the Behaviour: data/global.ex &
data/instance.ex.

The idea here is that we'll not only use these modules to define the
standard repo functionality, but also include our own functions which
manage the repos or migrate the databases.  This is done by database
type since they are really different in purpose and management.  I also
believe we get two supervision trees for each kind of data store, so
this is another advantage (I think).