Refactor msbms_startup_options.sample.toml
This refactor updates the startup options to include the idea of being
where the list of available servers is kept.  This is a bit different
than before, where we defined a special global server completely and
then defaults for all of the instances generically.

Now we have a list of one or more [db_server] tables that define all of
the connection properties for a given server.  This includes things like
default pool sizes, how many instance databases a server may support,
and whether or not the instances are considered development or
production.  The Global database is simply now a reference to one of
these configured servers, with a couple special parameters for how the
global database itself will be connected.

Finally, while the idea of db_server here is very closely related to
actual database servers, the same database server installation may be
referenced as multiple different servers in this file.  This could be
useful for example to sculpt subscription models in a way that maximizes
database server resources or protects certain classes of instance users.