WIP add of global repo and server related funcs
This module Msbms.System.Data.Global defines the repo for the global
database and defines the functions for starting up the global database
repo and bootstrapping it/migrating it.

At this stage the code here is very preliminary and was written prior to
changes that might be relevant have been made elsewhere (i.e. DbServer).
As such this is very much work in progress and not complete in any
sense.