Refactor constants out of config.exs
While I think it's still reasonable to have some of these constants
defined in config.exs, I think having them be defined on a module level
as actual constants for the module makes for better organization.

  - Remove the constants from the config.exs file
  - Add new module Msbms.System.Constants to define constants for the
    system module.
  - Add an interface to the Msbms.System public api to get the constants

It's unclear if the public API will be useful or not as most of the
constants do inside baseball for the System module itself... but let's
expose it here for now.