Refactor system constant names for consistency
Both the system_db_name and system_db_login constants were named
inconsistently with other references to the global database through the
system; there was a time that we weren't going to use "global", but
"system" to refer to the internal operating database and related code.
Now we've pretty firmly settled on the global terminology even though it
is overloaded in other parts of the application (hopefully in a context
clear way).

So update these constants to be consistent with where we've landed.