Add type Msbms.System.Types.DatastoreOptions
Datastores are an abstraction in the application which define the basics
of connecting to a database.  This includes the name of the database,
the actual database connections to be started to operate the datastore,
and default pool sizes for standard database connections.  This commit
creates a type in the application to handle this data.

I can already see the next item to refactor which is to bring the
connection pool information into the specific database connections as
well as making clear that the datastore abstraction includes all
database connections needed for the application to interact with a
specific database on the database server.