Refactor database role related names
In a number of places we refer to the standard database roles in either
the long form (i.e. app_user, api_admin) or their short form (appusr,
apiadm).  This is somewhat inconsistent however and it would be cleaner
to make it consistent.

This commit renames most instances of the long form of the standard
database role names to be the short form.  We do not do this for the
names in the msbms_startup_options.sample.toml file, just the internal
references in the application; this is because the longer names are
slightly clearer for end users needing to set values in the TOML file.