Fix an PrivilegedDatastore design issue
As originally built, the privileged datastore was only really expecting
to connect to the global database... which is fine expect that the
privileged datastore needs to connect to any database server to perform
administrative actions.  This means functions need to know which server
is connected and do so in a way that allows concurrent administrative
connections to multiple database servers.

  - Add a DbServer referencing parameter to functions so that the
    correct database server connection can be found.

  - Update get_datastore_id to generate the correct datastore ID for any
    DbServer.