Add Msbms.System.Data.Privileged
This module adds database access functions which are considered
privileged database operations and require use of the privileged
datastore.

  - Add function Privileged.db_roles_exist? to test if a list of
    database roles exist on the database server or not.

  - Add function Privileged.db_role_exists? to test if a specific
    database role exists on the database server.

In both cases it's not clear how useful these will be; at least as far
as being part of an exposed API.  It is clear something like these
functions will be useful for internal operations involving the creation
of database roles.