Refactor migration file management code in builder
Observed an issue where running the migration builder when migration
files already exist would cause the SQL to be appended to the migration
rather than overwriting the migration.

There is a question raised by this which is: should the builder actually
overwrite previously built migrations?  The best answer would be to do
so only in the presence of an option being passed explicitly requesting
a full rebuild.  But even that's not completely clear as then we'd have
to have some way to know what migrations which migrations were already
released and which should be rebuilt as part of ongoing development.

For the foreseeable future, we won't have any actually released software
and everything should be rebuilt when the migrations are built; even
when there are releases the original release files shouldn't change
anyway so any rebuild of released migrations should simply reconstruct
the original outcome.

So on rebuild:

  - At the start of the build process recursively delete the directory
    for the database type being created.

  - Explicitly create the database type directory, including any missing
    parents.

  - Remove the "touch" operations as creating the file happens
    implicitly when opening it for writing when it doesn't already
    exist.

  - Change the mode in which the file is opened from "append" to
    exclusive.  This will error if the file already exists, but it
    shouldn't exist since we deleted the directories/files in earlier
    steps of the build process.

  - Add an onscreen indication that a migration was created.  It's just
    to give the developer warm/fuzzy feelings that something expected
    actually happened.