Add a couple of salt values to Constants
While I'm not overfond of this whole password management thing (or using
passwords at all for some of these purposes), having components of these
things spread out into different at-rest origins seems reasonable
(though that opinion is fully unqualified).

This commit adds a couple salt values to the code as one of those
origins.  It should be noted that these values should never be used as a
credential on their own: they should always be combined with other
elements that are drawn from elsewhere.

(Really this feels more like the idea of "Pepper", but what we call it
doesn't matter so much).