Add generate_password function to Data.Utils
As stated elsewhere, using algorithmic password generation feels like a
bad idea, but for now it should be good enough.

This commit adds a feature to take the various components (all sourced
from differing at-rest persistence locations) and performs the operation
to create the database role passwords.

As always, this shouldn't be considered a robust security solution, but
simply a placeholder in lieu of a qualified review and solution for db
role authentication.