Refactor attrib naming in startup options TOML
The term "instance" in the naming of these attributes was misleading as
"instance" has a more precise meaning now.  So update these attribute
names to reference use "server" instead.

Note that there are a few places where this change is made that are not
in this commit, because they are bound up with other, more important
changes and will be committed with those changes.