Refactor "salt" to "code" in InstanceConfig
I think this attribute was refactored once to be salt from code and this
is reverting that change.  It's a more accurate name for what this
string really does.  (There's an argument that "salt" isn't  really an
accurate term at all... but that's a pedantic argument that we needn't
worry about for now).