Refactor nested function calls to be piped
In GlobalDatastore and InstanceDatastore, there were a couple instances
where we had nested function calls and it's really more stylistically
correct to use the pipe operator.  So do that, and also be sure that
the database_owner attribute in both modules are explicitly down-cased;
the database might not care, but matching in Elixir does care.