Refactor PrivilegedDatastore role checks for pid
The PrivilegedDatastore connections to the database were changed from
being managed on a named basis to using PIDs for reference.  This change
was made so that concurrent processes could have their own connections
if that scenario should ever arise.  Since privileged connections are
transient, being started as needed, it was conceivable that a privileged
process could finish and close the connection while another privileged
process was not yet finished.  Now each of those operations would have
their own database connection.

Note that this does mean that there's a chance of asking for too many
database connections now since we'll open as many as we need for the
number of requested privileged operations.  In practice though, this
should be a fairly low risk... something to watch and to be sure to
handle well in application code.