Refactor for new credential creations
This commit puts in place the new automatically generated credentials
into the GlobalDatastore and InstanceDatastore modules.  As said
elsewhere, this probably won't be good enough for secure role access in
the end, but will suffice for now and until there's sufficient reason to
get a qualified review of these role management mechanisms.