Fix db_role_exists? casing in Data.Privileged
The function that checks for a database role already existing was doing
a case sensitive check on data that was, due to the way SQL works, being
recorded as lowercase and thus the existence check failed.

Fix this by explicitly downcasing the rolename prior to doing the
lookup.  In fact, this downcasing is now redundant as we now explicitly
do this downcase at the time the role names are created, but this extra
step won't hurt now and may help in future.