Add Actions.Privileged and system_start function
This commit adds a new module Actions.Privileged which contains a
prototype of a system bootstrap function.  At this stage it can:

  - Check to see if the global database roles exist and create them if
    they don't

  - Check to see if the global database itself exists and create it if
    it doesn't exist yet

  - Check the current migration state of the global database and apply
    any outstanding migrations that have not been applied; this includes
    bootstrapping the global database and the database_migration table
    if not yet bootstrapped

This won't be the final form for this functionality, but it gets the
key pieces in place to finish up a working version.

Note that this module is a rename of the Msbms.System.Actions.Global
module.  Since the bootstrapping process is really privileged, it makes
more sense to call this global database bootstrapping and startup
privileged functionality.