Refactor out unneeded database access contexts
We're using database roles associated with certain generalized access
rights that we're calling access or security contexts.  These are
barriers at the database level for who can do what with data and the
database functional layer.  The idea is sound (I believe), but it turns
out that things are unlikely to be as fine-grained as I previously
believed.  The admin and user roles for the application are unlikely to
offer significant differences in access rights, and certainly not enough
to justify having extra database roles (and database connections
consumed), at least not for now.  This is especially true now that the
privileged and global datastores exist and are their own connections.

So, remove the "admin" related roles and keep the user related roles.
While I don't think there will be huge security differences between the
"app" and "api" flavors that we also divide by, we can scale the user
access and api access differently and make some finer grained
availability decisions with that bifurcation.

I think this will likely save about a 3rd of the database connections I
was previously planning on consuming per instance.