Rename Msbms.System.Actions.Privileged to Global
Even though we're depending on PrivilegedDatastore for most of the
actions, Msbms.Systsem.Actions.Global is still the better name.  This is
especially true given that we'll also have Instance actions requiring
PrivilegedDatastore and they'll look very similar to Global.  Having the
naming reflect the subject matter, rather that the required services to
perform the action, just seems more natural.