Refactor database location to standalone directory
As the application is moved to a more componentized architecture, the
database definition looks more like it's own module.  This is because
the global and instance logic will be split into multiple Elixir
applications and while the global/instance specific tables could also
live in those modules, their are shared database files which we'd have
to duplicate.

At some point we can refactor the build destinations to respect the
individual modules when running the database build script.