Add msbms_build elixir project
This elixir project will hold elixir based build scripts, such as
the builddb mix task.  The reality is that this should probably just be
a standalone script... but I don't how to correctly build or run those
yet so starting with a new elixir project since it will directly
translate more easily for now.