Refactor Elixir code locations
To better structure the elixir code, add some more organizing
directories.  Now we will put all the component Elixir projects in the
components directory, the application which assembles the components
into a working application in the application directory, interfaces into
the application (such as web UI and external API) into the application
directory, and finally the build tools into the tools directory.