Update msbms_syst_startup_options/mix.exs
This adds the dependencies in mix.lock (a result of mix.exs settings)
and establishes the formatting and dependencies of this module.

At some point we should try and find a well accepted, smaller library
for validations than Ecto Changesets.  While Ecto will do the job, it
just feels heavy.