Add component msbms_syst_error; replaces error.ex
This is a small error defining library which replaces the
Msbms.System.Types.Error module in msbms.

This looks fairly complete and provides a better error type definition
than the original.  The only downside is that the whole error stack
approach feels un-elixir like, but it may yet be worthwhile given the
kinds of functional stacking we see in large, cross-functional business
applications.

  - Remove the error.ex file (Msbms.System.Types.Error) from Elixir
    project msbms

  - Create new project msbms_syst_error to replace the deleted module

  - Add an error struct (MsbmsSystError.Impl.MsbmsError) to define what
    a structured error condition looks like.

  - Add gettext translation system and generate .pot file.  Still
    working through exactly how this works.  Getting the gettext() macro
    calls around the text is the most important thing now since figuring
    out where to put them later will be tricky (easier to remove later
    if not needed than to add later).

  - Add basic functions for getting user visible error descriptions and
    root cause from error stack.