Add component msbms_syst_options
The functionality here comes from the now removed
msbms_syst_startup_options component.  msbms_syst_options is a more
generalized way to access TOML encoded options files in the file system
and forward a map of those options to the callers.  This also means this
functionality is no longer necessarily for only the application
startup options, but for any options that might be encoded into a file
system based TOML file.

With this generalization, none of the value specific functionality, such
as dbserver handling, has been brought into this component.  All such
value specific handling should now happen in the modules that will make
use of the values retrieved from the TOML file.