Clean up dependencies
Reorder the dependencies list for ease of finding dependencies (true
mostly elsewhere) and to comply with msbms dependency order standards
that are generally applicable.

Also, allow the most recent non-breaking versions of dependencies to be
pulled on new builds.  There's a little speculation on the correct
version to test for depenedency to dependency, but there was some effort
put in to get it right.