Refactor application msbms Elixir project
The truth is that this is basically starting fresh from Elixir's mix new
msbms.  The original version of msbms was moved to msbms_old (not under
source control) and a new project created with the same name as the
original.

Since moving the Dave Thomas recommended component based architecture,
it seemed to make more sense to start fresh and to copy anything of
continuing relevance from the original project into this one.

This project will be the layer where the individual components are
integrated and is intended to be releasable via mix release.  There will
be some future thought to how best to support the user interface and API
projects, but that question doesn't need to be solved yet and we know
that some form of application only server should be supported. (should
help to decouple application logic from the user/integration
interfaces).