Fix GitHub Actions msbms_application.yml config
The original test used a pattern for running multiple tasks ("-run")
seen on the erlef/setup-beam repository.  But that pattern was not
compatible with our yml config and I didn't appreciate that until the
config failed to run.

Fix by moving the second run step into it's own named step.