Add application/msbms_old to .gitignore
This entry should only be temporary.  msbms_old is the original
monolithic Elixir project that was the start of development.  With the
advent of the component based architecture, a new msbms Elixir project
was started, but code in the original project files is still being
ported.

So, for now, exclude that directory in .gitignore, but it should be
removed once the msbms_old directory has outlived its usefulness and
been itself removed.