Fix credo detected errors in msbms_syst_options
There were two issues to fix.  1) Was a readability issue in a test
script where the aliases weren't listed in alphabetical order; and
2) an issue where a with expression was returning the same values as the
last function in the call which credo detected as redundant.

This commit fixes both issues and locally credo tests fine in this
component now.

Per CI testing