Update Github Actions for multiple jobs
Previously we had one big job with all of the components tested inside.
This update changes each component and the application to be its own CI
job.

This might take longer since it looks like it has to rebuild the
environment for each job, but I'm hoping a failure in any one job
doesn't prevent the other jobs from running; currently the first
error in a job stops everything cold at the point of the error.

Finally, this might might not be the best arrangement (by component).
If may be better to include all the compiling in one job, the testing in
another, and finally one for credo.  Hard to say.  But for now I think
we'll be OK.