Refactor get_options in msbms_syst_options
The "with" expression didn't feel right.  Refactor maybe_file_decode to
expect an error tuple from maybe_file_read and just pass it on if found.
This allows us to replace the with block in get_options with a simple
pipeline instead.

This change makes maybe_file_decode a bit more brittle in its use cases
as it's really tuned to the output of maybe_file_read now, but I think
that's OK since this is a small amount of functionality and we're
working with private functions.