Refactor msbms databases for datastore changes
A number of database related testing elements where included in the
global and instance database build plans.  Since msbms_syst_datastore is
self testing, we no longer need these testing pieces so they're removed
in this commit.

Also, the individual database definitions are no longer expected to
create their own database migrations management tables, so remove both
the common definition and the inclusion of that definition in the build
plans.