Add new func. type to enum_access_account_states
This commit adds the "pending" functional type to the available access
account states.  This state is intended to be used for new access
accounts where identity may need to be verified.  This would be the
case, for example, when independent bookkeepers have an un-owned access
account that they can use with multiple owners.