Add allow_global_logins to syst_access_accounts
This column is being added to facilitate the move of authentication from
being instance controlled to being a global, infrastructure level
concern.

When users that are broadly considered "staff" are allowed to log into
the system, they rightfully may have access to multiple instances or
instances owned by different companies (tenants).  However, if users are
acting in the capacity of "customer" or "vendor" accessing portal
functionality enabled by tenants, the system should be more transparent
and the user experience should be more direct with the specific tenant.

This implements that context for when selection of application instances
is appropriate at login time vs. when it's not appropriate.