Remove syst_identities identifier unique const.
When performing central but owner/instance directed authentication, it's
possible for different access accounts to be created for the same real
world account holder using a single identifier (email).  In this case,
we use references to the specific owner or instance to distinguish
between different access accounts... but this means that identifier
isn't unique for each identity type.

It must be unique for each owner, so a validation trigger is required to
prove validity under the new rules.